Renaming diff command classes.
