+++ liquibase-core/src/main/java/liquibase/command/core/DiffCommand.java
@@ -140,14 +140,14 @@ public class DiffCommand extends AbstractCommand<CommandResult> {
     @Override
     public CommandResult run() throws Exception {
         final CommandScope commandScope = new CommandScope("diffInternal");
-        commandScope.addArgumentValue(InternalDiffCommandStep.REFERENCE_DATABASE_ARG, this.referenceDatabase);
-        commandScope.addArgumentValue(InternalDiffCommandStep.TARGET_DATABASE_ARG, this.targetDatabase);
-        commandScope.addArgumentValue(InternalDiffCommandStep.SNAPSHOT_TYPES_ARG, this.snapshotTypes);
-        commandScope.addArgumentValue(InternalDiffCommandStep.SNAPSHOT_LISTENER_ARG, this.snapshotListener);
-        commandScope.addArgumentValue(InternalDiffCommandStep.REFERENCE_SNAPSHOT_CONTROL_ARG, this.referenceSnapshotControl);
-        commandScope.addArgumentValue(InternalDiffCommandStep.TARGET_SNAPSHOT_CONTROL_ARG, this.targetSnapshotControl);
-        commandScope.addArgumentValue(InternalDiffCommandStep.OBJECT_CHANGE_FILTER_ARG, this.objectChangeFilter);
-        commandScope.addArgumentValue(InternalDiffCommandStep.COMPARE_CONTROL_ARG, this.compareControl);
+        commandScope.addArgumentValue(DiffCommandStep.REFERENCE_DATABASE_ARG, this.referenceDatabase);
+        commandScope.addArgumentValue(DiffCommandStep.TARGET_DATABASE_ARG, this.targetDatabase);
+        commandScope.addArgumentValue(DiffCommandStep.SNAPSHOT_TYPES_ARG, this.snapshotTypes);
+        commandScope.addArgumentValue(DiffCommandStep.SNAPSHOT_LISTENER_ARG, this.snapshotListener);
+        commandScope.addArgumentValue(DiffCommandStep.REFERENCE_SNAPSHOT_CONTROL_ARG, this.referenceSnapshotControl);
+        commandScope.addArgumentValue(DiffCommandStep.TARGET_SNAPSHOT_CONTROL_ARG, this.targetSnapshotControl);
+        commandScope.addArgumentValue(DiffCommandStep.OBJECT_CHANGE_FILTER_ARG, this.objectChangeFilter);
+        commandScope.addArgumentValue(DiffCommandStep.COMPARE_CONTROL_ARG, this.compareControl);
 
         commandScope.setOutput(this.outputStream);
         commandScope.execute();
