+++ liquibase-integration-tests/src/test/groovy/liquibase/diffchangelog/DiffChangelogTest.groovy
@@ -4,7 +4,7 @@ import liquibase.Scope
 import liquibase.command.CommandResultsBuilder
 import liquibase.command.CommandScope
 import liquibase.command.core.InternalDiffChangelogCommandStep
-import liquibase.command.core.InternalDiffCommandStep
+import liquibase.command.core.DiffCommandStep
 import liquibase.database.Database
 import liquibase.database.DatabaseFactory
 import liquibase.diff.compare.CompareControl
@@ -47,10 +47,10 @@ CREATE TABLE $tableName ( product_no varchar(20) DEFAULT nextval('$sequenceName'
             DatabaseFactory.instance.openDatabase(postgres.getConnectionUrl().replace("lbcat", "lbcat2"), postgres.getUsername(), postgres.getPassword(), null, null)
 
         CommandScope commandScope = new CommandScope(InternalDiffChangelogCommandStep.COMMAND_NAME)
-        commandScope.addArgumentValue(InternalDiffCommandStep.REFERENCE_DATABASE_ARG, refDatabase)
+        commandScope.addArgumentValue(DiffCommandStep.REFERENCE_DATABASE_ARG, refDatabase)
         commandScope.addArgumentValue(InternalDiffChangelogCommandStep.CHANGELOG_FILE_ARG, changelogfile)
-        commandScope.addArgumentValue(InternalDiffCommandStep.TARGET_DATABASE_ARG, targetDatabase)
-        commandScope.addArgumentValue(InternalDiffCommandStep.COMPARE_CONTROL_ARG, CompareControl.STANDARD)
+        commandScope.addArgumentValue(DiffCommandStep.TARGET_DATABASE_ARG, targetDatabase)
+        commandScope.addArgumentValue(DiffCommandStep.COMPARE_CONTROL_ARG, CompareControl.STANDARD)
         commandScope.addArgumentValue(InternalDiffChangelogCommandStep.DIFF_OUTPUT_CONTROL_ARG,  new DiffOutputControl())
         OutputStream outputStream = new ByteArrayOutputStream()
         CommandResultsBuilder commandResultsBuilder = new CommandResultsBuilder(commandScope, outputStream)
@@ -94,10 +94,10 @@ COMMENT ON COLUMN $viewName.$columnName IS '$columnComment';
                 DatabaseFactory.instance.openDatabase(postgres.getConnectionUrl().replace("lbcat", "lbcat2"), postgres.getUsername(), postgres.getPassword(), null, null)
 
         CommandScope commandScope = new CommandScope(InternalDiffChangelogCommandStep.COMMAND_NAME)
-        commandScope.addArgumentValue(InternalDiffCommandStep.REFERENCE_DATABASE_ARG, refDatabase)
+        commandScope.addArgumentValue(DiffCommandStep.REFERENCE_DATABASE_ARG, refDatabase)
         commandScope.addArgumentValue(InternalDiffChangelogCommandStep.CHANGELOG_FILE_ARG, changelogfile)
-        commandScope.addArgumentValue(InternalDiffCommandStep.TARGET_DATABASE_ARG, targetDatabase)
-        commandScope.addArgumentValue(InternalDiffCommandStep.COMPARE_CONTROL_ARG, CompareControl.STANDARD)
+        commandScope.addArgumentValue(DiffCommandStep.TARGET_DATABASE_ARG, targetDatabase)
+        commandScope.addArgumentValue(DiffCommandStep.COMPARE_CONTROL_ARG, CompareControl.STANDARD)
         commandScope.addArgumentValue(InternalDiffChangelogCommandStep.DIFF_OUTPUT_CONTROL_ARG,  new DiffOutputControl())
         OutputStream outputStream = new ByteArrayOutputStream()
         CommandResultsBuilder commandResultsBuilder = new CommandResultsBuilder(commandScope, outputStream)
